Passed itinerary data to add item activity using a companion object.
Added an onclick listener which adds a step to an itinerary with the imputed data. I have it adding a step to the Paris itinerary for now.

There is a println statement in the listener which shows that the button is working correctly. When you first add a step, the println will print the name of that new step by accessing the itinerary. If you subsequently add more steps, it will still just print the original step added - showing that it is actually saving the data and not just printing what you just imputed.